requirements.txt
pyautogui>=0.9.54
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pytesseract>=0.3.10
pyserial>=3.5
requests>=2.31.0
mss>=9.0.0
